### 2017-08-02 #347 @whedon
[PRE REVIEW]: visdat: Visualising Whole Data Frames
**Submitting author:** @njtierney (<a href="http://orcid.org/0000-0003-1460-8722">Nicholas John Tierney</a>)
**Repository:** <a href="https://github.com/ropensci/visdat" target ="_blank">https://github.com/ropensci/visdat</a>
**Version:** v0.2.0
