### 2017-08-08 #347 @njtierney
Hi there,

So visdat has been through the rOpenSci onboarding process [here](https://github.com/ropensci/onboarding/issues/87). When I submitted the package to JOSS I linked to the package repo, not to the [Zenodo DOI](https://zenodo.org/record/837274), I'm wondering if perhaps this is why it didn't get processed the same as [rtimicropem](https://github.com/openjournals/joss-reviews/issues/333), perhaps @arfon can shed some light on this.

Additionally, I have made some minor changes in the paper.md document of visdat since the review of rOpenSci ended, it would be great if I could keep these changes, unless they are too much? You can see the changes [here](https://github.com/ropensci/visdat/commit/4d391b11a1adce8dca8b8d0efbe2f466cfeaa3dc)

This is my first time submitting something to JOSS, please let me know if I have missed anything, this seems like a great process! :)
