### 2017-08-10 #355 @arfon
Hi @njtierney, many thanks for your submission. This should be quick as you've already been through the rOpenSci process.

A couple of issues with your submission `paper.md`:

It looks like you've got a couple of different paper files, `paper.Rmd` and `paper.md`. We need the file to be named `paper.md` and for the YAML header to be formatted correctly as per [this example](https://github.com/arfon/fidgit/blob/master/paper/paper.md).

In addition, the markdown citations don't seem to be being picked up correctly (see the pandoc errors below)

```
pandoc-citeproc: reference dplyr not found
pandoc-citeproc: reference Rcore not found
pandoc-citeproc: reference ggplot2 not found
```

Finally, the current paper reads a little too much like technical documentation (i.e. naming methods etc) which is not the purpose of the JOSS paper (this material should be in the technical documentation). Please take a look at [this guidance](https://github.com/openjournals/joss/blob/487e1fe646ff0deb5a7f3b1649c3456ef0eafa64/app/views/home/about.html.erb#L73-L77) as to what your paper should contain 😄
