+++ paper/paper.md
@@ -1,4 +1,4 @@
-# visdat: Visualising whole data frames
+# visdat: Visualising Whole Data Frames
 01 August 2017  
 
 # Summary
@@ -48,15 +48,15 @@ dplyr::glimpse(typical_data)
 ```
 ## Observations: 5,000
 ## Variables: 9
-## $ ID         <chr> "0001", "0002", "0003", "0004", "0005", NA, "0007",...
-## $ Race       <fctr> White, Hispanic, White, Black, White, Hispanic, Hi...
-## $ Age        <int> 34, 25, 35, NA, NA, NA, 26, 31, 20, 26, NA, NA, 21,...
-## $ Sex        <fctr> Female, Male, Male, Female, Male, Male, Male, Fema...
-## $ Height(cm) <dbl> 181.6, 174.3, 171.9, 188.4, 171.1, 179.5, 175.8, 17...
-## $ IQ         <dbl> 100, 100, 104, 116, 106, 105, 95, 98, 88, NA, 96, 1...
-## $ Smokes     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FAL...
-## $ Income     <dbl> 33873.82, 38799.39, 7235.28, 29629.44, 19990.00, 69...
-## $ Died       <lgl> TRUE, TRUE, FALSE, TRUE, TRUE, FALSE, FALSE, TRUE, ...
+## $ ID         <chr> "0001", "0002", "0003", "0004", "0005", "0006", "00...
+## $ Race       <fctr> Black, Black, Black, Hispanic, NA, White, White, B...
+## $ Age        <chr> NA, "25", "31", "27", "21", "22", "23", "21", NA, "...
+## $ Sex        <fctr> Male, Male, Female, Female, Female, Female, Female...
+## $ Height(cm) <dbl> 175.9, 171.7, 173.5, 172.4, 158.5, 169.5, 163.7, 16...
+## $ IQ         <dbl> 110, 84, 115, 84, 116, 83, 101, 97, 92, 99, 88, 86,...
+## $ Smokes     <lgl> FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL...
+## $ Income     <fctr> 4334.29, 16682.37, 50402.01, 91442.86, 75266.05, 1...
+## $ Died       <lgl> FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL...
 ```
 
 Looking at this, you might then ask: "Isn't it odd that Income is a factor? And Age is a character?". And you might start to wonder what else is different, what else changed? It might be a bit unclear where to go from there. Do you do some exploratory plots of the data? What if the plot looks weird because the data has strange types? What if the data has other strange features? 
@@ -82,5 +82,9 @@ vis_miss(typical_data)
 ![](paper_files/figure-html/vis-miss-1.png)<!-- -->
 
 These functions provide useful tools to help "get a look at the data", using preliminary visualisation techniques. The plots are built using ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations. This means that users can customise and extend graphics from visdat very easily.
+
+# Acknowlegements
+
+I would like to thank the two reviewers, Mara Averick and Sean Hughes, for their helpful suggestions that resulted in a much better package.
 
 # References
