### 2017-08-11 #355 @njtierney
Hi @arfon !

OK sure thing, I'm reviewing these changes at the moment and will have something to you at the end of next week at the latest.

Question - is it OK if I use `paper.Rmd` (to generate a figure), which will generate `paper.md`, and then I edit that as per the guidelines?

Cheers! :)
