### 2017-08-11 #355 @arfon
> Question - is it OK if I use `paper.Rmd` (to generate a figure), which will generate `paper.md`, and then I edit that as per the guidelines?

I think that should be OK. Our automated submission tooling will pick up any file named `paper.md` so just make sure that is the one that has the JOSS paper in.
