### 2017-08-16 bb6ae84
update JOSS paper and adjust figures
