+++ paper/paper.Rmd
@@ -1,74 +1,51 @@
 ---
 title: 'visdat: Visualising Whole Data Frames'
-tags:
-  - visualisation
-  - R
-  - exporatory data analysis
 authors:
- - name: Nicholas Tierney
-   orcid: 0000-0003-1460-8722
-   affiliation: 1
-affiliations:
- - name: Monash University
-   index: 1
-date: 01 August 2017
-bibliography: paper.bib
-output: 
-  html_document: 
+- affiliation: 1
+  name: Nicholas Tierney
+  orcid: 0000-0003-1460-8722
+date: "01 August 2017"
+output:
+  pdf_document: default
+  html_document:
     keep_md: yes
+bibliography: paper.bib
+tags:
+- visualisation
+- R
+- exporatory data analysis
+affiliations:
+- index: 1
+  name: Monash University
 ---
+
+JOSS requirements:
+
+- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
+- A clear statement of need that illustrates the purpose of the software
+- A list of key references including a link to the software archive
+- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
 
 # Summary
 
-Reading in a new dataset means looking at the data to get a sense of what it contains, and potential problems and challenges with the data to get it analysis ready. "Looking at the data" can mean different things. Often times you look at the first six rows of data, the head of the data:
+When you receive a new dataset you need to look at the data to get a sense of what is in it and understand potential problems and challenges to get it analysis-ready. "Taking a look at the data" can mean different things. For example, examining statistical summaries (minimum, maximum, mean, interquartile range), finding missing values, checking data formatting, creating graphical summaries such as histograms, scatterplots, boxplots, and more.
 
-```{r head-iris}
+When handling typical "messy" data, these preliminary exploratory steps can become difficult to perform when you encounter data that is not what you expect. For example, income might be a factor instead of numeric, date could be a number not a character string or a date class, or values could be missing when they shouldn't be. Often times you realise that you have expectations of the data, but it can be hard to know what they are exactly until they are a problem. This is sort of like how you might not think to buy more lightbulbs until one goes out - when you go to use the data in an exploratory scatter plot, or in a preliminary model, you don't always realise your data is in the wrong format until that moment. What you need is a birds eye view of the data that tells you what classes are in the dataframe, and where the missing data are.
 
-head(iris)
+`visdat` is an R [@Rcore] package that provides a tool to "get a look at the data", creating heatmap-like visualisations of an entire dataframe - classes in the data, missing values, and comparisons between two datasets. `visdat` takes inspiration from [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), and is powered by ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations that can be extended if needed. 
 
-```
-
-Or you can `glimpse` the data, using the `dplyr` package [@dplyr]
-
-```{r glimpse}
-
-dplyr::glimpse(iris)
-
-```
-
-This shows that there are 150 observations, and 5 variables, which are doubles and a factor, along with some of the values in the data. However, we usually don't have data like the canonical iris dataset. Let's take a look at some data that might be a bit more typical of "messy" data.
-
-```{r visdat-glimpse}
-
-library(visdat)
-dplyr::glimpse(typical_data)
-
-```
-
-Looking at this, you might then ask: "Isn't it odd that Income is a factor? And Age is a character?". And you might start to wonder what else is different, what else changed? It might be a bit unclear where to go from there. Do you do some exploratory plots of the data? What if the plot looks weird because the data has strange types? What if the data has other strange features? 
-
[…]
