### 2017-08-16 f544450
Minor wording changes in JOSS paper
