+++ paper/paper.md
@@ -1,108 +1,35 @@
----
-title: 'visdat: Visualising Whole Data Frames'
-tags:
-  - visualisation
-  - R
-  - exporatory data analysis
-authors:
- - name: Nicholas Tierney
-   orcid: 0000-0003-1460-8722
-   affiliation: 1
-affiliations:
- - name: Monash University
-   index: 1
-date: 01 August 2017
-bibliography: paper.bib
----
-
-
-
 # visdat: Visualising Whole Data Frames
+01 August 2017  
+
+JOSS requirements:
+
+- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
+- A clear statement of need that illustrates the purpose of the software
+- A list of key references including a link to the software archive
+- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
 
 # Summary
 
-Reading in a new dataset means looking at the data to get a sense of what it contains, and potential problems and challenges with the data to get it analysis ready. "Looking at the data" can mean different things. Often times you look at the first six rows of data, the head of the data:
+When you receive a new dataset you need to look at the data to get a sense of what is in it and understand potential problems and challenges to get it analysis-ready. "Taking a look at the data" can mean different things. For example, examining statistical summaries (minimum, maximum, mean, inter-quartile range), finding missing values, checking data formatting, creating graphical summaries such as histograms, scatter plots, box plots, and more.
 
+When handling typical real-world data, these preliminary exploratory steps can become difficult to perform when values are not what you expect. For example, income might be a factor instead of numeric, date could be a number not a character string or a date class, or values could be missing when they shouldn't be. Often times you realise that you have expectations of the data, which are hard to realise until they are a problem. This is similar to how you might not think to buy more light bulbs until one goes out: when you use data in an exploratory scatter plot, or a preliminary model, you often don't realise your data is in the wrong format until that moment. What is needed is a birds eye view of the data, which tells you what classes are in the dataframe, and where the missing data are.
 
-```r
-head(iris)
-```
+`visdat` is an R [@Rcore] package that provides a tool to "get a look at the data", creating heatmap-like visualisations of an entire dataframe providing information on classes in the data, missing values, and also comparisons between two datasets. `visdat` takes inspiration from [`csv-fingerprint`](https://github.com/setosa/csv-fingerprint), and is powered by ggplot2 [@ggplot2], which provides a consistent and powerful framework for visualisations that can be extended if needed. 
 
-```
-##   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
-## 1          5.1         3.5          1.4         0.2  setosa
-## 2          4.9         3.0          1.4         0.2  setosa
-## 3          4.7         3.2          1.3         0.2  setosa
-## 4          4.6         3.1          1.5         0.2  setosa
-## 5          5.0         3.6          1.4         0.2  setosa
-## 6          5.4         3.9          1.7         0.4  setosa
-```
-
-Or you can `glimpse` the data, using the `dplyr` package [@dplyr]
-
-
-```r
-dplyr::glimpse(iris)
-```
-
-```
-## Observations: 150
-## Variables: 5
-## $ Sepal.Length <dbl> 5.1, 4.9, 4.7, 4.6, 5.0, 5.4, 4.6, 5.0, 4.4, 4.9,...
-## $ Sepal.Width  <dbl> 3.5, 3.0, 3.2, 3.1, 3.6, 3.9, 3.4, 3.4, 2.9, 3.1,...
[…]
