### 2017-08-17 ebf659f
update paper for JOSS submission
