+++ paper/paper.Rmd
@@ -19,13 +19,6 @@ affiliations:
   name: Monash University
 ---
 
-JOSS requirements:
-
-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
-- A clear statement of need that illustrates the purpose of the software
-- A list of key references including a link to the software archive
-- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
-
 # Summary
 
 When you receive a new dataset you need to look at the data to get a sense of what is in it, and understand potential problems and challenges to get it analysis-ready. "Taking a look at the data" can mean different things. For example: examining statistical summaries (minimum, maximum, mean, inter-quartile range), finding missing values, checking data formatting, creating graphical summaries such as histograms, scatter plots, box plots, and more.
