### 2017-08-17 #355 @njtierney
Hi @arfon !

I've submitted the changes for the [paper.md](https://github.com/ropensci/visdat/blob/master/paper/paper.md) - do you still get the same pandoc citation errors?
