### 2017-08-18 #355 @arfon
> I've submitted the changes for the paper.md - do you still get the same pandoc citation errors?

It looks good now, thanks!

At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
