### 2017-08-18 4e6a95f
updated links and bumped to version 0.2.1
