+++ man/vis_dat_ly.Rd
@@ -33,5 +33,6 @@ vis_dat_ly(airquality)
 
 }
 \seealso{
-\code{\link[=vis_dat]{vis_dat()}} \code{\link[=vis_miss]{vis_miss()}} \code{\link[=vis_miss_ly]{vis_miss_ly()}} \code{\link[=vis_guess]{vis_guess()}} \code{\link[=vis_compare]{vis_compare()}}
+\code{\link[=vis_dat]{vis_dat()}} \code{\link[=vis_miss]{vis_miss()}} \code{\link[=vis_miss_ly]{vis_miss_ly()}} \code{\link[=vis_guess]{vis_guess()}}
+\code{\link[=vis_compare]{vis_compare()}}
 }
