### 2017-08-21 d8da792
update DOI badges in README and other docs
