### 2017-08-21 #355 @njtierney
Yup!

Here is the DOI link: https://zenodo.org/badge/latestdoi/50553382

And the release: https://github.com/ropensci/visdat/releases/tag/v0.2.1

Is that the info that you need?
