### 2017-08-22 #355 @arfon
@njtierney - your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00355 ⚡️ 🚀 💥
